backend/requirements.txt
django==4.2.7
djangorestframework==3.14.0
django-filter==23.5
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0
groq==0.15.0